+++ packages/modul-components/src/utils/qa/components/elements/qa-elements.html
@@ -1,0 +1,29 @@
+<div class="m-qa-elements">
+    <h2>
+        Elements
+    </h2>
+    <ul>
+        <li v-for="(element, index) in elements"
+            :key="index"
+            style="display: flex; align-items: center; justify-content: space-between;">
+            <div @mouseenter="onElementMouseEnter(element.id)"
+                 @mouseleave="onElementMouseLeave(element.id)"
+                 @click="$emit('element-selected', element)"
+                 style="display: flex; align-items: center; cursor: pointer;">
+                {{element.id}}
+            </div>
+            <div>
+                <m-icon name="m-svg__file-text"
+                        color="blue"></m-icon>
+                <m-icon name="m-svg__error"
+                        color="red"></m-icon>
+                <m-icon v-if="!elementHasError(element)"
+                        name="m-svg__clock"
+                        color="orange"></m-icon>
+                <m-icon v-if="!elementHasError(element)"
+                        name="m-svg__completed-filled"
+                        color="green"></m-icon>
+            </div>
+        </li>
+    </ul>
+</div>
