+++ packages/modul-components/src/utils/qa/components/elements/qa-elements.scss
@@ -1,0 +1,3 @@
+.m-qa-elements {
+
+}
