+++ packages/modul-components/src/utils/qa/components/login/qa-login.html
@@ -1,12 +1,7 @@
 <div class="m-qa-login">
-    <div style="display: flex; align-items: center; justify-content: space-between;">
-        <span class="m-u--h2">
-            QA-Plugin
-        </span>
-        <img height="40"
-             src="https://dataverse.scholarsportal.info/logos/75125/UL_C.png"
-             alt="">
-    </div>
+    <h2>
+        QA-Plugin
+    </h2>
     <m-form :form-group="formGroup"
             @submit="submit">
         <m-input-group class="m-u--margin-top--l">
