+++ packages/modul-components/src/utils/qa/components/logs/qa-logs.html
@@ -1,0 +1,5 @@
+<div class="m-qa-logs">
+    <h2>
+        Logs
+    </h2>
+</div>
