+++ packages/modul-components/src/utils/qa/components/logs/qa-logs.scss
@@ -1,0 +1,3 @@
+.m-qa-logs {
+
+}
