+++ packages/modul-components/src/utils/qa/qa-panel.html
@@ -1,43 +1,12 @@
-<div :class="inputClass"
-     style="z-index: 2000;">
+<div class="m-qa-panel">
     <div v-if="expanded">
         <m-qa-login v-if="!user"></m-qa-login>
         <div v-else>
-            <div v-if="!selectedElement">
-                <ul>
-                    <li v-for="(element, index) in elements"
-                        :key="index"
-                        style="display: flex; align-items: center; justify-content: space-between;">
-                        <div @mouseenter="onElementMouseEnter(element.id)"
-                             @mouseleave="onElementMouseLeave(element.id)"
-                             @click="selectedElement = element"
-                             style="display: flex; align-items: center; cursor: pointer;">
-                            {{element.id}}
-                        </div>
-                        <div>
-                            <m-icon name="m-svg__file-text"
-                                    color="blue"></m-icon>
-                            <m-icon name="m-svg__error"
-                                    color="red"></m-icon>
-                            <m-icon v-if="!elementHasError(element)"
-                                    name="m-svg__clock"
-                                    color="orange"></m-icon>
-                            <m-icon v-if="!elementHasError(element)"
-                                    name="m-svg__completed-filled"
-                                    color="green"></m-icon>
-                        </div>
-                    </li>
-                </ul>
-            </div>
-            <div v-else>
-                {{selectedElement.id}}
-                <ul>
-                    <li v-for="(log, index) in selectedElement.logs"
-                        :key="index + 'log'">
-                        {{log.body}}
-                    </li>
-                </ul>
-            </div>
+            <m-qa-elements v-if="!selectedElement"
+                           :elements="elements"
+                           @element-selected="selectedElement = $event"></m-qa-elements>
+            <m-qa-logs v-else
+                       :element="selectedElement"></m-qa-logs>
         </div>
     </div>
 </div>
